==> R/02_clean.qmd <==
---
title: "Clean"
format: html
editor: visual
---

## Loading libraries

```{r}
library(tidyverse)
library(dplyr)
```

## First clean: make it tidy

In the first clean, the quotation marks have been removed and each value has been given its own cell.

First we load in the functions

```{r}
source("99_proj_func.R")
```

### Cleaning 1. data part

```{r}
# Running on the first part of data (mean)
cancer_data_1 <- cleaning_data("/path/to/R/data_1.csv")
```

### Cleaning 2. data part

```{r}
# Running on second part of data (SE)
cancer_data_2 <- cleaning_data("/path/to/R/data_2.csv")

```

### Cleaning 3. data part

```{r}
# Running on third part of data (Worst)
cancer_data_3 <- cleaning_data("/path/to/R/data_3.csv")
```

### Joining the data parts into one dataset

```{r}
output_path = "/path/to/02_dat_clean.tsv" # Where output should be added

output_path_binary = "/path/to/02_dat_clean_binary.tsv" # Where output should be added

# Joining the 3 data parts
cancer_dataset_clean <- cancer_data_1 |> 
  full_join(cancer_data_2, by = 'id') |> 
  full_join(cancer_data_3, by = 'id')|>
  as_tibble() |>
  write_tsv(file = output_path) |>
  mutate(diagnosis = case_when(diagnosis == "M" ~ 0,
    diagnosis == "B" ~ 1)) |> 
  write_tsv(file = output_path_binary) # Making dataset with binary diagnosis (1 = Benign, 0 = Malignant)


```
