==> R/03_describe.qmd <==
---
title: "Describe data"
format: html
editor: visual
  markdown: 
    wrap: 72
---

## Loading libraries

```{r}
library(tidyverse)
library(dplyr)
library(ggplot2)
```

## Loading data

```{r}
cancer_dataset_clean <- read_tsv("/path/to/data/R/02_dat_clean.tsv")

```

## Heatmap of correlation between parameters

```{r}
# Specify the folder path and filename for the saved plot
folder_path <- "/path/to/results"
```

```{r}
file_name <- "Heatmap.png"

# Creating correlation matrix
correlation_matrix <- cancer_dataset_clean |> 
  select(!c(diagnosis, id)) |> # removes id and non-numeric values
  cor() # do correlaton on data

myrownames <- rownames(correlation_matrix) # saves row names

# Rearrange data structure
tidy_correlation <- correlation_matrix |>
  as_tibble() |> # makes correlation matrix into tibble
  mutate(myrownames) |> # add the row names
  pivot_longer(cols = contains(c('mean', 'worst', 'se')), # rearrange data
               names_to = "variables",
               values_to = "correlation")

# Plotting heatmap
ggplot_heatmap <- ggplot(data = tidy_correlation,
                         mapping = aes(x = variables,
                                       y = myrownames,
                                       fill = correlation)) +
  geom_tile(color = "white") +
  scale_fill_gradient2(low = "blue", 
                       mid = "white", 
                       high = "red", 
                       midpoint = 0, 
                       limits = c(-1, 1)) +
  theme_minimal() + 
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  labs(title = "Correlation between variables",
       x = "variables",
       y = "variables") +
  theme(plot.title = element_text(hjust = 0.5))

print(ggplot_heatmap)

# Save the ggplot as a PNG file in the specified folder
ggsave(file.path(folder_path, file_name), ggplot_heatmap, width = 8, height = 6, dpi = 300)
```

## Boxplot over data

Loading in function

```{r}
source('../R/99_proj_func.R')
```

Making boxplot

```{r}
# Name of specific figure
file_name <- "boxplot_mean.png"


# Gather the mean values using function from 99_proj_func.R
pivot_longer_mean <- pivot_standardized('mean', cancer_dataset_clean)


# Customize boxplot
gg <- ggplot(data = pivot_longer_mean, 
       aes(x = Parameter, y = StandardizedValue, fill = factor(diagnosis))) +
  
  geom_boxplot(alpha = 0.8, 
               color = "black", 
               size = 0.2) +
  
    scale_fill_manual(values = c("M" = "#Ff1000", "B" = "#0012B2"),
                    labels = c("malignant", "benign")) + 
  
  labs(y = "Standardized Value", 
       title = "Boxplot of parameters (mean)",
       fill = "diagnosis") +
  
  theme_minimal() +
  theme(
    plot.margin = margin(2, 1, 1, 1, "cm"),
    axis.text.x = element_text(angle = 45, hjust = 1)
  )

# Save the ggplot as a PNG file in the specified folder
ggsave(file.path(folder_path, file_name), gg, width = 8, height = 6, dpi = 300)

```

```{r}
# Name of specific figure
file_name <- "boxplot_se.png"

# set seed
set.seed(1234)

# Gather the mean values
pivot_longer_se <- pivot_standardized('se', cancer_dataset_clean)


# Customize boxplot
gg <- ggplot(data = pivot_longer_se, 
       aes(x = Parameter, y = StandardizedValue, fill = factor(diagnosis))) +
  
  geom_boxplot(alpha = 0.8, 
               color = "black", 
               size = 0.2) +
  
  scale_fill_manual(values = c("M" = "#Ff1000", "B" = "#0012B2"),
                    labels = c("malignant", "benign")) + 
  
  labs(y = "Standardized Value", 
       title = "Boxplot of parameters (mean)",
       fill = "diagnosis") +
  
  labs(x = "Parameter", y = "Standardized Value", title = "Boxplot of parameters (se)") +
  theme_minimal() +
  theme(
    plot.margin = margin(2, 1, 1, 1, "cm"),
    axis.text.x = element_text(angle = 45, hjust = 1)
  )
# Save the ggplot as a PNG file in the specified folder
ggsave(file.path(folder_path, file_name), gg, width = 8, height = 6, dpi = 300)


```

```{r}
# Name of specific figure
file_name <- "boxplot_worst.png"

# set seed
set.seed(1234)

# Gather the mean values
pivot_longer_worst <- pivot_standardized('worst', cancer_dataset_clean)


# Customize boxplot
gg <- ggplot(data = pivot_longer_worst, 
       aes(x = Parameter, y = StandardizedValue, fill = factor(diagnosis))) +
  
  geom_boxplot(alpha = 0.8, 
               color = "black", 
               size = 0.2) +
  
  scale_fill_manual(values = c("M" = "#Ff1000", "B" = "#0012B2"),
                    labels = c("malignant", "benign")) + 
  
  labs(y = "Standardized Value", 
       title = "Boxplot of parameters (mean)",
       fill = "diagnosis") +
  
  labs(x = "Parameter", y = "Standardized Value", title = "Boxplot of parameters (worst)") +
  theme_minimal() +
  theme(
    plot.margin = margin(2, 1, 1, 1, "cm"),
    axis.text.x = element_text(angle = 45, hjust = 1)
  )

# Save the ggplot as a PNG file in the specified folder
ggsave(file.path(folder_path, file_name), gg, width = 8, height = 6, dpi = 300)



```
