==> R/05_PCA_analysis_01.qmd <==
---
title: "PCA"
format: html
editor: visual
---

## PCA

## Loading libraries

```{r}
library(tidyverse)
library(broom)  
library(cowplot)
```

## Read data

```{r}
cancer_dataset_clean <- read_tsv("/net/pupil1/home/people/s195796/projects/Group_10_project/R/02_dat_clean.tsv")
```

## Making Principal component analysis

```{r}
cancer_dataset_clean <- cancer_dataset_clean |>
  select(!id)

pca_fit <- cancer_dataset_clean |> 
  select(!diagnosis) |> # retain only numeric columns
  prcomp(scale = TRUE) # do PCA on scaled data

```

## Plotting PCA

```{r}
# Specify the folder path and filename for the saved plot
folder_path <- "/net/pupil1/home/people/s195796/projects/Project_group_10/results"
```

```{r}

file_name <- "PCA.png"

PCA_plot <- pca_fit |>
  augment(cancer_dataset_clean) |> # add original dataset back in
  ggplot(aes(.fittedPC1, .fittedPC2, color = diagnosis)) + 
  geom_point(size = 1.5) +
  scale_color_manual(
    values = c(M = "#Ff1000", B = "#0012B2")
  ) +
  theme_half_open(12) + 
  background_grid() +
  theme_minimal_grid(12) +
    labs(title = "PCA plot",
         x = "PC1",
         y = "PC2"
       ) +
  theme(plot.title = element_text(hjust = 0.5))

# Save the ggplot as a PNG file in the specified folder
ggsave(file.path(folder_path, file_name), PCA_plot, width = 8, height = 6, dpi = 300)
```

## Looking at the data in PC coordinates

```{r}
# extract rotation matrix
pca_fit |>
  tidy(matrix = "rotation")
```

```{r}
# define arrow style for plotting
arrow_style <- arrow(
  angle = 20, ends = "first", type = "closed", length = grid::unit(8, "pt")
)

# plot rotation matrix
pca_fit |>
  tidy(matrix = "rotation") |>
  pivot_wider(names_from = "PC", names_prefix = "PC", values_from = "value") |>
  ggplot(aes(PC1, PC2)) +
  geom_segment(xend = 0, yend = 0, arrow = arrow_style) +
  geom_text(
    aes(label = column),
    hjust = 1, 
    nudge_x = -0.2,
    color = "#300e9d",
    size = 2
  ) +
  xlim(-1.25, .5) + ylim(-.5, 1) +
  coord_fixed() + # fix aspect ratio to 1:1
  theme_minimal_grid(12) +
    labs(title = "Rotation matrix",
       ) +
  theme(plot.title = element_text(hjust = 0.5))
```

The rotation matrix is implicitly applied to the data points to align them with the principal components. The resulting plot shows the data points projected onto the new coordinate system defined by the principal components.

## Looking at the variance explained by each PC

```{r}
pca_fit |>
  tidy(matrix = "eigenvalues")
```

## Plotting variation in data shown by the PC

```{r}

file_name <- "Variation_in_data.png"

PCA_variation <- pca_fit |>
  tidy(matrix = "eigenvalues") |>
  ggplot(aes(PC, percent)) +
  geom_col(fill = "#300e9d", alpha = 0.8) +
  scale_x_continuous(breaks = 1:9) +
  scale_y_continuous(
    labels = scales::percent_format(),
    expand = expansion(mult = c(0, 0.01))
  ) +
  theme_minimal_hgrid(12) +
  labs(title = "Variation in data",
       ) +
  theme(plot.title = element_text(hjust = 0.5))

# Save the ggplot as a PNG file in the specified folder
ggsave(file.path(folder_path, file_name), PCA_variation, width = 8, height = 6, dpi = 300)
```

The first component captures more than 40% of the variation in the data
